chore: move strapi to lib folder
