+++ src/lib/strapi/__mocks__/fetch-all-articles-projection.json
