+++ src/lib/strapi/__mocks__/fetch-all-tags-query-response.json
