+++ src/lib/strapi/__mocks__/fetch-config-response.json
