+++ src/lib/strapi/__mocks__/fetch-home-page-response.json
