+++ src/lib/strapi/__mocks__/fetch-privacy-policy-page-response.json
