+++ src/lib/strapi/mappers.test.ts
@@ -2,7 +2,7 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
-import { DEFAULT_SITE_TITLE } from "../../../constants/defaults";
+import { DEFAULT_SITE_TITLE } from "../../constants/defaults";
 import {
 	mapArticleToDetail,
 	mapArticleToSummary,
