+++ src/lib/strapi/queries/about-and-config.ts
